Separate out ApiServer class #51
